move index.ts to allo.ts
